+++ bootmenu/script/2nd-init.sh
@@ -1,4 +1,5 @@
-#!/system/xbin/sh
+#!/sbin/busybox ash
+
 ######## BootMenu Script
 ######## Execute [2nd-init] Menu
 
@@ -6,31 +7,27 @@ source /system/bootmenu/script/_config.sh
 
 ######## Main Script
 
-mount -o remount,rw /
+toolbox mount -o remount,rw rootfs /
 rm -f /*.rc
 rm -f /*.sh
-rm -rf /osh
+rm -f /osh
 rm -rf /preinstall
 cp -f /system/bootmenu/2nd-init/* /
+chmod 640 /*.rc
+chmod 750 /init
+rm -f /sbin/ueventd
 ln -s /init /sbin/ueventd
-cp -f /system/bin/adbd /sbin/adbd
+cp -f /system/bootmenu/binary/adbd /sbin/adbd
+killall ueventd
 
 ADBD_RUNNING=`ps | grep adbd | grep -v grep`
 if [ -z "$ADB_RUNNING" ]; then
-    rm -f /sbin/adbd.root
     rm -f /tmp/usbd_current_state
-    #delete if is a symlink
-    [ -L "/tmp" ] && rm -f /tmp
-    mkdir -p /tmp
-else
-    # well, not beautiful but do the work
-    # to keep current usbd state
-    if [ -L "/tmp" ]; then
-        mv /tmp/usbd_current_state / 2>/dev/null
-        rm -f /tmp
-        mkdir -p /tmp
-        mv /usbd_current_state /tmp/ 2>/dev/null
-    fi
+fi
+
+# original /tmp data symlink
+if [ -L /tmp.bak ]; then
+  rm /tmp.bak
 fi
 
 if [ -L /sdcard-ext ]; then
@@ -47,21 +44,19 @@ umount /mnt/asec
 umount /mnt/obb
 umount /cache
 umount /data
-mount -o remount,rw,relatime,mode=775,size=128k /dev
 
 ######## Cleanup
 
 rm -f /sbin/lsof
 
 ## busybox cleanup..
-for cmd in $(/sbin/busybox --list); do
-    [ -L "/sbin/$cmd" ] && rm "/sbin/$cmd"
-done
+if [ -f /sbin/busybox ]; then
+    for cmd in $(/sbin/busybox --list); do
+        [ -L "/sbin/$cmd" ] && rm "/sbin/$cmd"
+    done
 
-rm -f /sbin/busybox
-
-## used for adbd shell (can be bash also)
-/system/xbin/ln -s /system/xbin/busybox /sbin/sh
+    rm -f /sbin/busybox
+fi
 
 ## reduce lcd backlight to save battery
 echo 18 > /sys/class/leds/lcd-backlight/brightness
